Open profile edit, view needs editing
